# openapi_dump.json
{"openapi": "3.1.0", "info": {"title": "Student Transcript Grade Pipeline API", "description": "API \u0111\u1ecdc b\u1ea3ng \u0111i\u1ec3m h\u1ecdc sinh b\u1eb1ng Ollama Vision v\u00e0 l\u1ecdc m\u00f4n kh\u00f4ng \u0111\u1ea1t k\u1ebft h\u1ee3p OpenCV HSV.", "version": "1.0.0"}, "paths": {"/auth/login": {"post": {"tags": ["auth"], "summary": "Login", "operationId": "login_auth_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginRequest"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/auth/me": {"get": {"tags": ["auth"], "summary": "Read Users Me", "operationId": "read_users_me_auth_me_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/documents": {"get": {"tags": ["documents"], "summary": "Get Documents", "operationId": "get_documents_documents_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"OAuth2PasswordBearer": []}]}, "post": {"tags": ["documents"], "summary": "Upload Document", "operationId": "upload_document_documents_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_upload_document_documents_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"OAuth2PasswordBearer": []}]}}, "/documents/{doc_id}": {"get": {"tags": ["documents"], "summary": "Get Document", "operationId": "get_document_documents__doc_id__get", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Doc Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["documents"], "summary": "Update Document", "operationId": "update_document_documents__doc_id__put", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Doc Id"}}], "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_update_document_documents__doc_id__put"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["documents"], "summary": "Delete Document", "operationId": "delete_document_documents__doc_id__delete", "security": [{"OAuth2PasswordBearer": []}], "parameters": [{"name": "doc_id", "in": "path", "required": true, "schema": {"type": "integer", "title": "Doc Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/health": {"get": {"summary": "Check Health", "description": "Ki\u1ec3m tra tr\u1ea1ng th\u00e1i server v\u00e0 k\u1ebft n\u1ed1i t\u1edbi Ollama.\nL\u1ea5y danh s\u00e1ch c\u00e1c model \u0111\u00e3 c\u00e0i \u0111\u1eb7t tr\u00ean m\u00e1y.", "operationId": "check_health_api_health_get", "parameters": [{"name": "host", "in": "query", "required": false, "schema": {"type": "string", "default": "http://localhost:11434", "title": "Host"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/process": {"post": {"summary": "Process Transcript", "description": "Ti\u1ebfp nh\u1eadn file b\u1ea3ng \u0111i\u1ec3m (\u1ea3nh/PDF), ch\u1ea1y to\u00e0n b\u1ed9 pipeline v\u00e0 tr\u1ea3 v\u1ec1 k\u1ebft qu\u1ea3.", "operationId": "process_transcript_api_process_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_process_transcript_api_process_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/certificate/process": {"post": {"summary": "Process Certificate", "description": "Ti\u1ebfp nh\u1eadn \u1ea3nh ch\u1ee9ng ch\u1ec9 (JPG/PNG), tr\u00edch xu\u1ea5t th\u00f4ng tin qua Ollama Vision\nv\u00e0 t\u1ef1 \u0111\u1ed9ng so kh\u1edbp v\u1edbi danh m\u1ee5c \u0111i\u1ec1u ki\u1ec7n t\u1ed1t nghi\u1ec7p b\u1eb1ng rapidfuzz.", "operationId": "process_certificate_api_certificate_process_post", "requestBody": {"content": {"multipart/form-data": {"schema": {"$ref": "#/components/schemas/Body_process_certificate_api_certificate_process_post"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/api/download/{file_type}": {"get": {"summary": "Download Result File", "description": "T\u1ea3i v\u1ec1 file k\u1ebft qu\u1ea3 CSV ho\u1eb7c JSON theo session_id.", "operationId": "download_result_file_api_download__file_type__get", "parameters": [{"name": "file_type", "in": "path", "required": true, "schema": {"type": "string", "title": "File Type"}}, {"name": "session_id", "in": "query", "required": true, "schema": {"type": "string", "title": "Session Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}}, "components": {"schemas": {"Body_process_certificate_api_certificate_process_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}, "model": {"anyOf": [{"type": "string"}, {"type": "null"}], "title": "Model"}, "host": {"type": "string", "title": "Host", "default": "http://localhost:11434"}}, "type": "object", "required": ["file"], "title": "Body_process_certificate_api_certificate_process_post"}, "Body_process_transcript_api_process_post": {"properties": {"file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}, "model": {"type": "string", "title": "Model", "default": "llama3.2-vision"}, "host": {"type": "string", "title": "Host", "default": "http://localhost:11434"}}, "type": "object", "required": ["file"], "title": "Body_process_transcript_api_process_post"}, "Body_update_document_documents__doc_id__put": {"properties": {"title": {"type": "string", "title": "Title"}, "file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "title": "Body_update_document_documents__doc_id__put"}, "Body_upload_document_documents_post": {"properties": {"title": {"type": "string", "title": "Title"}, "file": {"type": "string", "contentMediaType": "application/octet-stream", "title": "File"}}, "type": "object", "required": ["title", "file"], "title": "Body_upload_document_documents_post"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginRequest": {"properties": {"username": {"type": "string", "title": "Username"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["username", "password"], "title": "LoginRequest"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}, "input": {"title": "Input"}, "ctx": {"type": "object", "title": "Context"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"OAuth2PasswordBearer": {"type": "oauth2", "flows": {"password": {"scopes": {}, "tokenUrl": "/auth/login"}}}}}}